refactor: Version page better imports
