+++ package.json
@@ -16,24 +16,24 @@
     "@material-ui/core": "3.9.3",
     "@material-ui/icons": "3.0.2",
     "@octokit/rest": "16.28.7",
-    "@testing-library/react": "9.1.0",
+    "@testing-library/react": "9.1.3",
     "@types/enzyme": "3.10.3",
-    "@types/lodash": "4.14.136",
+    "@types/lodash": "4.14.137",
     "@types/material-ui": "0.21.6",
-    "@types/node": "12.7.1",
-    "@types/react": "16.9.1",
-    "@types/react-dom": "16.8.5",
-    "@types/react-router-dom": "4.3.4",
-    "@types/validator": "10.11.2",
+    "@types/node": "12.7.2",
+    "@types/react": "16.9.2",
+    "@types/react-dom": "16.9.0",
+    "@types/react-router-dom": "4.3.5",
+    "@types/validator": "10.11.3",
     "@verdaccio/babel-preset": "2.0.0",
     "@verdaccio/eslint-config": "2.0.0",
-    "@verdaccio/types": "8.0.0-next.2",
+    "@verdaccio/types": "8.0.0",
     "autosuggest-highlight": "3.1.1",
     "babel-loader": "8.0.6",
     "bundlesize": "0.18.0",
     "codeceptjs": "2.2.1",
     "codecov": "3.5.0",
-    "concurrently": "4.1.1",
+    "concurrently": "4.1.2",
     "cross-env": "5.2.0",
     "css-loader": "3.2.0",
     "date-fns": "1.30.1",
@@ -56,11 +56,11 @@
     "husky": "3.0.3",
     "identity-obj-proxy": "3.0.0",
     "in-publish": "2.0.0",
-    "jest": "24.8.0",
+    "jest": "24.9.0",
     "jest-emotion": "10.0.14",
-    "jest-environment-jsdom": "24.8.0",
+    "jest-environment-jsdom": "24.9.0",
     "jest-environment-jsdom-global": "1.2.0",
-    "jest-environment-node": "24.8.0",
+    "jest-environment-node": "24.9.0",
     "js-base64": "2.5.1",
     "js-yaml": "3.13.1",
     "localstorage-memory": "1.0.3",
@@ -76,7 +76,7 @@
     "react-autosuggest": "9.4.3",
     "react-dom": "16.9.0",
     "react-emotion": "9.2.12",
-    "react-hot-loader": "4.12.10",
+    "react-hot-loader": "4.12.11",
     "react-router": "5.0.1",
     "react-router-dom": "5.0.1",
     "resolve-url-loader": "3.1.0",
@@ -98,10 +98,10 @@
     "verdaccio": "4.2.1",
     "verdaccio-auth-memory": "1.1.5",
     "verdaccio-memory": "2.0.0",
-    "webpack": "4.39.1",
+    "webpack": "4.39.2",
     "webpack-bundle-analyzer": "3.4.1",
     "webpack-bundle-size-analyzer": "3.0.0",
-    "webpack-cli": "3.3.6",
+    "webpack-cli": "3.3.7",
     "webpack-dev-server": "3.8.0",
     "webpack-merge": "4.2.1",
     "whatwg-fetch": "3.0.0",
