+++ src/components/Author/Author.test.tsx
@@ -20,7 +20,7 @@ describe('<Author /> component', () => {
       },
     };
 
-    jest.doMock('../../pages/version/Version', () => ({
+    jest.doMock('../../pages/Version', () => ({
       DetailContextConsumer: component => {
         return component.children({ packageMeta });
       },
@@ -39,7 +39,7 @@ describe('<Author /> component', () => {
       },
     };
 
-    jest.doMock('../../pages/version/Version', () => ({
+    jest.doMock('../../pages/Version', () => ({
       DetailContextConsumer: component => {
         return component.children({ packageMeta });
       },
@@ -63,7 +63,7 @@ describe('<Author /> component', () => {
       },
     };
 
-    jest.doMock('../../pages/version/Version', () => ({
+    jest.doMock('../../pages/Version', () => ({
       DetailContextConsumer: component => {
         return component.children({ packageMeta });
       },
