+++ src/pages/Version/index.ts
@@ -1,0 +1,3 @@
+export { DetailContextProps, VersionPageConsumerProps } from './types';
+export { DetailContext, DetailContextConsumer, DetailContextProvider } from './context';
+export { default } from './Version';
