--- src/pages/version/index.ts
@@ -1,2 +1,0 @@
-export { DetailContextProps } from './types';
-export { default } from './Version';
